Rename plugin to `ClientForwardRefsPlugin`
