+++ ariadne_codegen/contrib/__init__.py
@@ -1,5 +1,5 @@
+from .client_forward_refs import ClientForwardRefsPlugin
 from .extract_operations import ExtractOperationsPlugin
-from .no_global_imports import NoGlobalImportsPlugin
 from .no_reimports import NoReimportsPlugin
 from .shorter_results import ShorterResultsPlugin
 
@@ -7,5 +7,5 @@ __all__ = [
     "ExtractOperationsPlugin",
     "NoReimportsPlugin",
     "ShorterResultsPlugin",
-    "NoGlobalImportsPlugin",
+    "ClientForwardRefsPlugin",
 ]
