+++ tests/main/clients/client_forward_refs/custom_scalars.py
