+++ tests/main/clients/client_forward_refs/expected_client/client.py
