+++ tests/main/clients/client_forward_refs/expected_client/client_forward_refs_fragments.py
