+++ tests/main/clients/client_forward_refs/expected_client/get_animal_by_name.py
