+++ tests/main/clients/client_forward_refs/expected_client/get_animal_fragment_with_extra.py
@@ -2,7 +2,7 @@ from typing import List
 
 from pydantic import Field
 
-from .no_global_imports_fragments import ListAnimalsFragment
+from .client_forward_refs_fragments import ListAnimalsFragment
 
 
 class GetAnimalFragmentWithExtra(ListAnimalsFragment):
