+++ tests/main/clients/client_forward_refs/expected_client/get_authenticated_user.py
