+++ tests/main/clients/client_forward_refs/expected_client/list_animals.py
