+++ tests/main/clients/client_forward_refs/expected_client/list_strings_3.py
