+++ tests/main/clients/client_forward_refs/expected_client/unwrap_fragment.py
@@ -1,0 +1,5 @@
+from .client_forward_refs_fragments import FragmentWithSingleField
+
+
+class UnwrapFragment(FragmentWithSingleField):
+    pass
