+++ tests/main/clients/client_forward_refs/pyproject.toml
@@ -2,10 +2,10 @@
 schema_path = "schema.graphql"
 queries_path = "queries.graphql"
 include_comments = "none"
-target_package_name = "no_global_imports"
+target_package_name = "client_forward_refs"
 files_to_include = ["custom_scalars.py"]
-fragments_module_name = "no_global_imports_fragments"
-plugins = ["ariadne_codegen.contrib.no_global_imports.NoGlobalImportsPlugin"]
+fragments_module_name = "client_forward_refs_fragments"
+plugins = ["ariadne_codegen.contrib.client_forward_refs.ClientForwardRefsPlugin"]
 
 [tool.ariadne-codegen.scalars.SimpleScalar]
 type = ".custom_scalars.SimpleScalar"
