+++ tests/main/clients/client_forward_refs/queries.graphql
