+++ tests/main/clients/client_forward_refs/schema.graphql
