+++ tests/main/clients/client_forward_refs_shorter_results/custom_scalars.py
