+++ tests/main/clients/client_forward_refs_shorter_results/expected_client/get_animal_by_name.py
