+++ tests/main/clients/client_forward_refs_shorter_results/expected_client/get_complex_scalar.py
