+++ tests/main/clients/client_forward_refs_shorter_results/expected_client/list_strings_2.py
