+++ tests/main/clients/client_forward_refs_shorter_results/expected_client/subscribe_strings.py
