+++ tests/main/clients/client_forward_refs_shorter_results/expected_client/unwrap_fragment.py
@@ -1,0 +1,5 @@
+from .client_forward_refs_shorter_resultsfragments import FragmentWithSingleField
+
+
+class UnwrapFragment(FragmentWithSingleField):
+    pass
