+++ tests/main/clients/client_forward_refs_shorter_results/schema.graphql
