--- tests/main/clients/no_global_imports/expected_client/unwrap_fragment.py
@@ -1,5 +1,0 @@
-from .no_global_imports_fragments import FragmentWithSingleField
-
-
-class UnwrapFragment(FragmentWithSingleField):
-    pass
